BackEnd/requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
sqlalchemy==2.0.15
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==40.0.2
bcrypt==4.0.1
PyJWT==2.7.0
jwt==1.3.1